backend/requirements.txt
Django==3.2.25
djangorestframework==3.12.4
pandas==1.5.3
django-cors-headers==3.13.0
reportlab==3.6.13